extract prepareListContext helper so explain reuses listRows setup
