+++ src/Appwrite/Platform/Modules/Databases/Http/Databases/Collections/Documents/Action.php
@@ -8,10 +8,15 @@ use Appwrite\Event\Publisher\Func as FunctionPublisher;
 use Appwrite\Extend\Exception;
 use Appwrite\Functions\EventProcessor;
 use Appwrite\Platform\Modules\Databases\Http\Databases\Action as DatabasesAction;
+use Appwrite\Utopia\Database\Documents\User;
 use Appwrite\Utopia\Database\Validator\CustomId;
 use Utopia\Database\Database;
 use Utopia\Database\Document;
+use Utopia\Database\Exception\NotFound as NotFoundException;
+use Utopia\Database\Exception\Query as QueryException;
+use Utopia\Database\Query;
 use Utopia\Database\Validator\Authorization;
+use Utopia\Database\Validator\Query\Cursor;
 
 abstract class Action extends DatabasesAction
 {
@@ -503,5 +508,105 @@ abstract class Action extends DatabasesAction
         $queueForEvents->reset();
         $queueForRealtime->reset();
         $queueForWebhooks->reset();
+    }
+
+    /**
+     * Shared setup for any list-style read on the documents/rows API surface.
+     *
+     * Both listRows/listDocuments and explainRows/explainDocuments need the
+     * same auth checks, database+collection lookup, query parse, cursor
+     * resolution, and find-closure construction. Centralising it here is the
+     * single source of truth so the explain endpoint stays byte-identical to
+     * the real read it's explaining.
+     *
+     * Returned bundle:
+     *   'database'           Document  the user-facing database doc
+     *   'collection'         Document  the user-facing collection doc
+     *   'dbForDatabases'     Database  the per-database adapter the read runs against
+     *   'queries'            Query[]   parsed Query objects (cursor value resolved)
+     *   'collectionTableId'  string    physical table id (`database_X_collection_Y`)
+     *   'hasSelects'         bool      whether the queries include any select
+     *   'find'               callable  closure that runs the find — wraps in skipRelationships when there are no related selects
+     *
+     * @param  array<string>  $queries  raw stringified queries from the HTTP request
+     * @return array<string, mixed>
+     */
+    protected function prepareListContext(
+        string $databaseId,
+        string $collectionId,
+        array $queries,
+        Database $dbForProject,
+        User $user,
+        callable $getDatabasesDB,
+        Authorization $authorization,
+    ): array {
+        $isAPIKey = $user->isApp($authorization->getRoles());
+        $isPrivilegedUser = $user->isPrivileged($authorization->getRoles());
+
+        $database = $authorization->skip(fn () => $dbForProject->getDocument('databases', $databaseId));
+        if ($database->isEmpty() || (! $database->getAttribute('enabled', false) && ! $isAPIKey && ! $isPrivilegedUser)) {
+            throw new Exception(Exception::DATABASE_NOT_FOUND, params: [$databaseId]);
+        }
+
+        $collection = $authorization->skip(fn () => $dbForProject->getDocument('database_' . $database->getSequence(), $collectionId));
+        if ($collection->isEmpty() || (! $collection->getAttribute('enabled', false) && ! $isAPIKey && ! $isPrivilegedUser)) {
+            throw new Exception($this->getParentNotFoundException(), params: [$collectionId]);
+        }
+
+        try {
+            $queries = Query::parseQueries($queries);
+        } catch (QueryException $e) {
+            throw new Exception(Exception::GENERAL_QUERY_INVALID, $e->getMessage());
+        }
+
+        $dbForDatabases = $getDatabasesDB($database);
+
+        $cursor = Query::getCursorQueries($queries, false);
+        $cursor = \reset($cursor);
+
+        if ($cursor !== false) {
+            $validator = new Cursor;
+            if (! $validator->isValid($cursor)) {
+                throw new Exception(Exception::GENERAL_QUERY_INVALID, $validator->getDescription());
+            }
+
+            $documentId = $cursor->getValue();
+
+            try {
+                $cursorDocument = $authorization->skip(fn () => $dbForDatabases->getDocument('database_' . $database->getSequence() . '_collection_' . $collection->getSequence(), $documentId));
+            } catch (NotFoundException) {
+                // The collection metadata document exists but the backing
+                // store has no table for it. Treat as collection not-found
+                // so the caller sees a 404 instead of a 500.
+                throw new Exception($this->getParentNotFoundException(), params: [$collectionId]);
+            }
+
+            if ($cursorDocument->isEmpty()) {
+                $type = ucfirst($this->getContext());
+                throw new Exception(Exception::GENERAL_CURSOR_NOT_FOUND, "$type '{$documentId}' for the 'cursor' value not found.");
+            }
+
+            $cursor->setValue($cursorDocument);
+        }
+
+        $collectionTableId = 'database_' . $database->getSequence() . '_collection_' . $collection->getSequence();
+        $hasSelects = ! empty(Query::groupByType($queries)['selections']);
+
+        // When there are no select queries, relationship loading is skipped on
+        // the underlying find() to avoid pulling related documents the caller
+        // did not ask for.
+        $find = $hasSelects
+            ? fn () => $dbForDatabases->find($collectionTableId, $queries)
+            : fn () => $dbForDatabases->skipRelationships(fn () => $dbForDatabases->find($collectionTableId, $queries));
+
+        return [
+            'database' => $database,
+            'collection' => $collection,
+            'dbForDatabases' => $dbForDatabases,
+            'queries' => $queries,
+            'collectionTableId' => $collectionTableId,
+            'hasSelects' => $hasSelects,
+            'find' => $find,
+        ];
     }
 }
